requirements.txt
Flask==3.1.1
mysql-connector-python==8.2.0
requests==2.32.3
python-dotenv==1.0.0